Fix all warnings in CoreFoundation
-Werror is now enabled so there won't be any more in the future.

Places I noticed that had stubs which didn't have it now print a
message indicating that a stub was called.

Fix all CFNumber.c warnings

Fix CFNumberFormatter.c warnings

More warning fixes

Fix warnings

Bridged class warnings

Fix NSDate and NSDateComponents warnings

Warnings fixes

Warnings fixes

No NSSet and NSSharedKeyDictionary warnings

NSSharedKeySet and NSString warnings

Fix NSTimeZone and NSTimer warnings

Fix NSURL.m warnings

Remove NSLog uses

Remove references to Foundation class

Fix 64 bit warning

Fix format strings for NSArray

Fix all warnings

All warnings enabled

+++ NSFastEnumerationEnumerator.m
@@ -43,7 +43,8 @@
 
     if (state->mutationsPtr == NULL || *state->mutationsPtr != state->extra[0])
     {
-        [NSException raise:NSGenericException format:@"*** Collection <%s: 0x%x> was mutated while being enumerated", object_getClassName(_obj), _obj];
+        [NSException raise:NSGenericException format:@"*** Collection <%s: 0x%x> was mutated while being enumerated", 
+		object_getClassName(_obj), (unsigned int)_obj];
         return nil;
     }
